Wrapper for long
PyCrypto asserts for long on Python 2, so we have to really use them

+++ paramiko/primes.py
@@ -86,7 +86,7 @@ class ModulusPack (object):
         tries = int(tries)
         size = int(size)
         generator = int(generator)
-        modulus = long(modulus, 16)
+        modulus = int(modulus, 16)
 
         # weed out primes that aren't at least:
         # type 2 (meets basic structural requirements)
